Remove action-utils.js (and move functions into actions.js)
Cleans up utils.js further
