+++ webofneeds/won-owner-webapp/src/main/webapp/app/redux/state-store.js
@@ -3,7 +3,7 @@ import Immutable from "immutable";
 import { actionTypes } from "../actions/actions.js";
 import * as atomUtils from "./utils/atom-utils.js";
 import { parseMetaAtom } from "../reducers/atom-reducer/parse-atom.js";
-import { is, urisToLookupMap } from "../utils.js";
+import { is } from "../utils.js";
 import won from "../won-es6";
 import * as connectionUtils from "./utils/connection-utils.js";
 
@@ -261,4 +261,70 @@ function fetchOwnedAtomAndDispatch(atomUri, dispatch) {
       });
       return;
     });
+}
+
+/**
+ * Takes a single uri or an array of uris, performs the lookup function on each
+ * of them seperately, collects the results and builds an map/object
+ * with the uris as keys and the results as values.
+ * If any call to the asyncLookupFunction fails, the corresponding
+ * key-value-pair will not be contained in the result.
+ * @param uris
+ * @param asyncLookupFunction
+ * @param excludeUris uris to exclude from lookup
+ * @param abortOnError -> abort the whole crawl by breaking the promisechain instead of ignoring the failures
+ * @return {*}
+ */
+function urisToLookupMap(
+  uris,
+  asyncLookupFunction,
+  excludeUris = [],
+  abortOnError = false
+) {
+  //make sure we have an array and not a single uri.
+  const urisAsArray = is("Array", uris) ? uris : [uris];
+  const excludeUrisAsArray = is("Array", excludeUris)
+    ? excludeUris
+    : [excludeUris];
+
+  const urisAsArrayWithoutExcludes = urisAsArray.filter(uri => {
+    const exclude = excludeUrisAsArray.indexOf(uri) < 0;
+    if (exclude) {
+      return true;
+    } else {
+      return false;
+    }
+  });
+
+  const asyncLookups = urisAsArrayWithoutExcludes.map(uri =>
+    asyncLookupFunction(uri).catch(error => {
+      if (abortOnError) {
+        throw new Error(error);
+      } else {
+        console.error(
+          `failed lookup for ${uri} in utils.js:urisToLookupMap ` +
+            error.message,
+          "\n\n",
+          error.stack,
+          "\n\n",
+          urisAsArrayWithoutExcludes,
+          "\n\n",
+          uris,
+          "\n\n",
+          error
+        );
+        return undefined;
+      }
+    })
+  );
+  return Promise.all(asyncLookups).then(dataObjects => {
+    const lookupMap = {};
+    //make sure there's the same
+    uris.forEach((uri, i) => {
+      if (dataObjects[i]) {
+        lookupMap[uri] = dataObjects[i];
+      }
+    });
+    return lookupMap;
+  });
 }
