+++ webofneeds/won-owner-webapp/src/main/webapp/app/utils.js
@@ -53,106 +53,6 @@ export function deepFreeze(obj) {
   return Object.freeze(obj);
 }
 
-/*
- * @param obj an object-tree.
- *
- * @param prefix add a custom prefix to all generated constants.
- *
- * @returns a tree using the same structure as `o` but with
- *          all leaves being strings equal to their lookup path.
- * e.g.:
- * tree2constants({foo: null}) -> {foo: 'foo'}
- * tree2constants{{foo: {bar: null}}) -> {foo: {bar: 'foo.bar'}}
- * tree2constants{foo: null}, 'pfx') -> {foo: 'pfx.foo'}
- */
-export function tree2constants(obj, prefix = "") {
-  //wrap prefix in array
-  prefix = prefix === "" ? [] : [prefix];
-
-  return deepFreeze(
-    reduceAndMapTreeKeys(
-      (acc, k) => acc.concat(k),
-      acc => acc.join("."),
-      prefix,
-      obj
-    )
-  );
-}
-
-/**
- * Traverses down an object, reducing the keys with the reducer
- * and then applying the mapper once it reaches the leaves.
- * The function doesn't modify the input-object.
- * @param obj
- * @param acc the initial accumulator
- * @param reducer (acc, key) => newAcc
- * @param mapper (acc) => newAcc
- * @returns {*}
- */
-
-export function reduceAndMapTreeKeys(reducer, mapper, acc, obj) {
-  if (typeof obj === "object" && obj !== null) {
-    const accObj = {};
-    for (let k of Object.keys(obj)) {
-      accObj[k] = reduceAndMapTreeKeys(
-        reducer,
-        mapper,
-        reducer(acc, k),
-        obj[k]
-      );
-    }
-    return accObj;
-  } else {
-    return mapper(acc);
-  }
-}
-
-/**
- * Traverses an object-tree and produces an object
- * that is just one level deep but concatenating the
- * traversal path.
- *
- * ```
- * flattenTree({
- *   myInt: 1,
- *   myObj: {
- *      myProp: 2,
- *      myStr: 'asdf',
- *      foo: {
- *        bar: 3
- *      }
- *   }
- * });
- * // result:
- * // {
- * //   'myInt': 1,
- * //   'myObj__myProp' : 2,
- * //   'myObj__myStr' : 'asdf',
- * //   'myObj__foo__bar' : 3
- * // }
- * ```
- *
- * @param tree {object} the object-tree
- * @param delimiter {string} will be used to join the path. by default `__`
- * @returns {object} the flattened object
- */
-export function flattenTree(tree, delimiter = "__") {
-  const accObj = {}; //the accumulator accObject
-  function _flattenTree(node, pathAcc = []) {
-    for (let k of Object.keys(node)) {
-      const pathAccUpd = pathAcc.concat(k);
-      if (typeof node[k] === "object" && node[k] !== null) {
-        _flattenTree(node[k], pathAccUpd);
-      } else {
-        const propertyName = pathAccUpd.join(delimiter);
-        accObj[propertyName] = node[k];
-      }
-    }
-  }
-  _flattenTree(tree);
-  return accObj;
-}
-
 export function delay(milliseconds) {
   return new Promise(resolve =>
     window.setTimeout(() => resolve(), milliseconds)
@@ -168,7 +68,7 @@ export function delay(milliseconds) {
  * @param callback {function}
  * @return {function} the unsubscribe function generated by `subscribe`
  */
-export function watch(subscribe, select, callback) {
+function watch(subscribe, select, callback) {
   let unsubscribe = null;
 
   /*
@@ -206,19 +106,14 @@ export function watchImmutableRdxState(redux, path, callback) {
  */
 export function generateIdString(length) {
   const characters = "abcdefghijklmnopqrstuvwxyz0123456789";
+
+  const arrayOfRandoms = length =>
+    Array.apply(null, Array(length)).map(() => Math.random());
+
   return arrayOfRandoms(length)
     .map(randomFloat => Math.floor(randomFloat * characters.length))
     .map(randomPosition => characters.charAt(randomPosition))
     .join("");
-}
-
-/**
- * Generate array of random numbers.
- * @param length
- * @returns {*}
- */
-function arrayOfRandoms(length) {
-  return Array.apply(null, Array(length)).map(() => Math.random());
 }
 
 export function readAsDataURL(file) {
@@ -229,152 +124,6 @@ export function readAsDataURL(file) {
     reader.onerror = () => reject("Something exploded in the FileReader");
 
     reader.readAsDataURL(file);
-  });
-}
-
-/**
- * taken from: https://esdiscuss.org/topic/es6-iteration-over-object-values
- *
- * example usage:
- *
- * ```javascript
- * for (let [key, value] of entries(o)) {
- *   console.log(key, ' --> ', value)
- * }
- * ```
- * @param obj the object to generate a (key,value)-pair iterator for
- */
-export function* entries(obj) {
-  for (let key of Object.keys(obj)) {
-    yield [key, obj[key]];
-  }
-}
-
-/**
- * @param listOfLists e.g. [ [1,2], [3], [], [3,4,5] ]
- * @return {*} e.g. [1,2,3,3,4,5]
- */
-export function flatten(listOfLists) {
-  return listOfLists.reduce(
-    (flattendList, innerList) =>
-      innerList ? flattendList.concat(innerList) : [], //not concatenating `undefined`s
-    [] //concat onto empty list as start
-  );
-}
-
-/**
- * Takes a single uri or an array of uris, performs the lookup function on each
- * of them seperately, collects the results and builds an map/object
- * with the uris as keys and the results as values.
- * If any call to the asyncLookupFunction fails, the corresponding
- * key-value-pair will not be contained in the result.
- * @param uris
- * @param asyncLookupFunction
- * @param excludeUris uris to exclude from lookup
- * @param abortOnError -> abort the whole crawl by breaking the promisechain instead of ignoring the failures
- * @return {*}
- */
-export function urisToLookupMap(
-  uris,
-  asyncLookupFunction,
-  excludeUris = [],
-  abortOnError = false
-) {
-  //make sure we have an array and not a single uri.
-  const urisAsArray = is("Array", uris) ? uris : [uris];
-  const excludeUrisAsArray = is("Array", excludeUris)
-    ? excludeUris
-    : [excludeUris];
-
-  const urisAsArrayWithoutExcludes = urisAsArray.filter(uri => {
-    const exclude = excludeUrisAsArray.indexOf(uri) < 0;
-    if (exclude) {
-      return true;
-    } else {
-      return false;
-    }
-  });
-
-  const asyncLookups = urisAsArrayWithoutExcludes.map(uri =>
-    asyncLookupFunction(uri).catch(error => {
-      if (abortOnError) {
-        throw new Error(error);
-      } else {
-        console.error(
-          `failed lookup for ${uri} in utils.js:urisToLookupMap ` +
-            error.message,
-          "\n\n",
-          error.stack,
-          "\n\n",
-          urisAsArrayWithoutExcludes,
-          "\n\n",
-          uris,
-          "\n\n",
-          error
-        );
-        return undefined;
-      }
-    })
-  );
-  return Promise.all(asyncLookups).then(dataObjects => {
-    const lookupMap = {};
-    //make sure there's the same
-    uris.forEach((uri, i) => {
-      if (dataObjects[i]) {
-        lookupMap[uri] = dataObjects[i];
-      }
-    });
-    return lookupMap;
-  });
-}
-
-/**
- * Takes a single uri or an array of uris, performs the lookup function on each
- * of them seperately, collects the results and builds an map/object
- * with the uris as keys and the results as values.
- * If any call to the asyncLookupFunction fails, the corresponding
- * key-value-pair will not be contained in the success-result but rather in the failed-results.
- * @param uris
- * @param asyncLookupFunction
- * @param excludeUris uris to exclude from lookup
- * @return {*}
- */
-export function urisToLookupSuccessAndFailedMap(
-  uris,
-  asyncLookupFunction,
-  excludeUris = []
-) {
-  //make sure we have an array and not a single uri.
-  const urisAsArray = is("Array", uris) ? uris : [uris];
-  const excludeUrisAsArray = is("Array", excludeUris)
-    ? excludeUris
-    : [excludeUris];
-
-  const urisAsArrayWithoutExcludes = urisAsArray.filter(uri => {
-    const exclude = excludeUrisAsArray.indexOf(uri) < 0;
-    if (exclude) {
-      return true;
-    } else {
-      return false;
-    }
-  });
-
-  const asyncLookups = urisAsArrayWithoutExcludes.map(uri =>
-    asyncLookupFunction(uri).catch(error => {
-      return error;
-    })
-  );
-  return Promise.all(asyncLookups).then(dataObjects => {
-    const lookupMap = { success: {}, failed: {} };
-    //make sure there's the same
-    uris.forEach((uri, i) => {
-      if (dataObjects[i] instanceof Error) {
-        lookupMap["failed"][uri] = dataObjects[i];
-      } else if (dataObjects[i]) {
-        lookupMap["success"][uri] = dataObjects[i];
-      }
-    });
-    return lookupMap;
   });
 }
